mq: WIP output formatting
